requirements.txt
# Telegram Bot API wrapper
python-telegram-bot

# YouTube transcript API for fetching video captions
youtube-transcript-api

# OpenAI API client for GPT and Whisper models
openai==1.50.2

# Environment variable management
python-dotenv==1.0.1

# Asyncio for asynchronous processing
asyncio==3.4.3